+++ MoMoney/Converters/ShowValuePercentConverter.cs
@@ -1,15 +1,16 @@
-﻿using System.Globalization;
+﻿using MoMoney.Helpers;
+using System.Globalization;
 
 namespace MoMoney.Converters;
 
 /// <summary>
-/// Converts value to "*.**%" if Constants.ShowValue is true
+/// Converts value to "*.**%" if Utilities.ShowValue is true
 /// </summary>
 class ShowValuePercentConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
-        return Constants.ShowValue ? value : "*.**%";
+        return Utilities.ShowValue ? value : "*.**%";
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)