+++ MoMoney/Data/MoMoneydb.cs
@@ -1,5 +1,6 @@
 ﻿using SQLite;
 using MoMoney.Models;
+using MoMoney.Helpers;
 
 namespace MoMoney.Data;