+++ MoMoney/ViewModels/EditTransactionViewModel.cs
@@ -6,6 +6,7 @@ using MoMoney.Views;
 using MoMoney.Models;
 using MoMoney.Services;
 using MoMoney.Exceptions;
+using MoMoney.Helpers;
 
 namespace MoMoney.ViewModels;