+++ MoMoney/Views/EditTransactionPage.xaml.cs
@@ -1,3 +1,4 @@
+using MoMoney.Helpers;
 using MoMoney.ViewModels;
 
 namespace MoMoney.Views;