+++ MoMoney/Views/Settings/LoggingPage.xaml.cs
@@ -10,4 +10,14 @@ public partial class LoggingPage : ContentPage
 		BindingContext = vm;
 		Loaded += vm.Init;
 	}
+
+	/// <summary>
+	/// Updates the DataTemplateSelector
+	/// </summary>
+	/// <param name="sender"></param>
+	/// <param name="e"></param>
+    private void switch_Toggled(object sender, ToggledEventArgs e)
+    {
+		listview.RefreshView();
+    }
 }